refactor(modules): simplify OSMFeature class in separating its admin and API classes
