+++ src/urban_mapper/modules/urban_layer/urban_layers/osm_features.py
@@ -4,83 +4,39 @@ from pathlib import Path
 from beartype import beartype
 
 from urban_mapper.utils import require_attributes_not_none
+
+from .admin_features_ import AdminFeatures
 from ..abc_urban_layer import UrbanLayerBase
 
-import osmnx as ox
 from shapely.geometry import Polygon, MultiPolygon
-
-
-@beartype
-class FeatureNetwork:
-    def __init__(self) -> None:
-        self._features: gpd.GeoDataFrame | None = None
-
-    def load(self, method: str, tags: Dict[str, str | bool], **kwargs) -> None:
-        method = method.lower()
-        valid_methods = {"address", "bbox", "place", "point", "polygon"}
-        if method not in valid_methods:
-            raise ValueError(f"Invalid method. Choose from {valid_methods}")
-
-        if method == "address":
-            if "address" not in kwargs or "dist" not in kwargs:
-                raise ValueError("Method 'address' requires 'address' and 'dist'")
-            self._features = ox.features_from_address(
-                kwargs["address"], tags, kwargs["dist"]
-            )
-        elif method == "bbox":
-            if "bbox" not in kwargs:
-                raise ValueError("Method 'bbox' requires 'bbox'")
-            bbox = kwargs["bbox"]
-            if not isinstance(bbox, tuple) or len(bbox) != 4:
-                raise ValueError("'bbox' must be a tuple of (left, bottom, right, top)")
-            self._features = ox.features_from_bbox(bbox, tags)
-        elif method == "place":
-            if "query" not in kwargs:
-                raise ValueError("Method 'place' requires 'query'")
-            self._features = ox.features_from_place(kwargs["query"], tags)
-        elif method == "point":
-            if "center_point" not in kwargs or "dist" not in kwargs:
-                raise ValueError("Method 'point' requires 'center_point' and 'dist'")
-            self._features = ox.features_from_point(
-                kwargs["center_point"], tags, kwargs["dist"]
-            )
-        elif method == "polygon":
-            if "polygon" not in kwargs:
-                raise ValueError("Method 'polygon' requires 'polygon'")
-            polygon = kwargs["polygon"]
-            if not isinstance(polygon, (Polygon, MultiPolygon)):
-                raise ValueError("'polygon' must be a shapely Polygon or MultiPolygon")
-            self._features = ox.features_from_polygon(polygon, tags)
-
-    @property
-    def features(self) -> gpd.GeoDataFrame:
-        if self._features is None:
-            raise ValueError("Features not loaded. Call load() first.")
-        return self._features
 
 
 @beartype
 class OSMFeatures(UrbanLayerBase):
     def __init__(self) -> None:
         super().__init__()
-        self.feature_network: FeatureNetwork | None = None
+        self.feature_network: AdminFeatures | None = None
         self.tags: Dict[str, str] | None = None
 
     def from_place(
-        self, place_name: str, tags: Dict[str, str | bool], **kwargs
+        self, place_name: str, tags: Dict[str, str | bool | dict | list], **kwargs
     ) -> None:
         self.tags = tags
-        self.feature_network = FeatureNetwork()
+        self.feature_network = AdminFeatures()
         self.feature_network.load("place", tags, query=place_name, **kwargs)
         self.layer = self.feature_network.features.to_crs(
             self.coordinate_reference_system
         )
 
     def from_address(
-        self, address: str, tags: Dict[str, str | bool], dist: float, **kwargs
+        self,
+        address: str,
+        tags: Dict[str, str | bool | dict | list],
+        dist: float,
+        **kwargs,
     ) -> None:
         self.tags = tags
-        self.feature_network = FeatureNetwork()
+        self.feature_network = AdminFeatures()
         self.feature_network.load("address", tags, address=address, dist=dist, **kwargs)
         self.layer = self.feature_network.features.to_crs(
             self.coordinate_reference_system
@@ -89,11 +45,11 @@ class OSMFeatures(UrbanLayerBase):
     def from_bbox(
         self,
         bbox: Tuple[float, float, float, float],
-        tags: Dict[str, str | bool],
+        tags: Dict[str, str | bool | dict | list],
         **kwargs,
     ) -> None:
         self.tags = tags
-        self.feature_network = FeatureNetwork()
+        self.feature_network = AdminFeatures()
         self.feature_network.load("bbox", tags, bbox=bbox, **kwargs)
         self.layer = self.feature_network.features.to_crs(
             self.coordinate_reference_system
@@ -102,12 +58,12 @@ class OSMFeatures(UrbanLayerBase):
     def from_point(
         self,
         center_point: Tuple[float, float],
-        tags: Dict[str, str | bool],
+        tags: Dict[str, str | bool | dict | list],
         dist: float,
         **kwargs,
     ) -> None:
         self.tags = tags
-        self.feature_network = FeatureNetwork()
+        self.feature_network = AdminFeatures()
         self.feature_network.load(
             "point", tags, center_point=center_point, dist=dist, **kwargs
         )
@@ -116,10 +72,13 @@ class OSMFeatures(UrbanLayerBase):
         )
 
     def from_polygon(
-        self, polygon: Polygon | MultiPolygon, tags: Dict[str, str | bool], **kwargs
+        self,
+        polygon: Polygon | MultiPolygon,
+        tags: Dict[str, str | bool | dict | list],
+        **kwargs,
     ) -> None:
         self.tags = tags
-        self.feature_network = FeatureNetwork()
+        self.feature_network = AdminFeatures()
         self.feature_network.load("polygon", tags, polygon=polygon, **kwargs)
         self.layer = self.feature_network.features.to_crs(
             self.coordinate_reference_system
@@ -170,7 +129,8 @@ class OSMFeatures(UrbanLayerBase):
         )
         mapped_data[output_column] = mapped_data[unique_id]
         return self.layer, mapped_data.drop(
-            columns=[unique_id, "distance_to_feature", "index_right"]
+            columns=[unique_id, "distance_to_feature", "index_right"],
+            errors="ignore",
         )
 
     @require_attributes_not_none(
